refactor: relocate common classes to version levels
